# Manually fuse reindexing intermediates with blockwise reduction for cohorts.